Code cleanup 🧹
