+++ src/ignitus-Contributors/constants.js
@@ -1,1 +1,11 @@
 export const API_URL = 'https://ignitusrestapi.herokuapp.com/testimonials';
+export const UCB = 'https://storage.googleapis.com/ignitus_assets/ig-university-logo/universityLogos_ucbLogo.png';
+export const STANFORD = 'https://storage.googleapis.com/ignitus_assets/ig-university-logo/universityLogos_stanfordLogo.png';
+export const CMU = 'https://storage.googleapis.com/ignitus_assets/ig-university-logo/universityLogos_cmuLogo.png';
+export const HARVARD = 'https://storage.googleapis.com/ignitus_assets/ig-university-logo/universityLogos_harvardLogo.png';
+export const MIT = 'https://storage.googleapis.com/ignitus_assets/ig-university-logo/universityLogos_mitLogo.png';
+export const OXFORD = 'https://storage.googleapis.com/ignitus_assets/ig-university-logo/universityLogos_oxfordLogo.png';
+export const PRINCETON = 'https://storage.googleapis.com/ignitus_assets/ig-university-logo/universityLogos_princetonLogo.png';
+export const YALE = 'https://storage.googleapis.com/ignitus_assets/ig-university-logo/universityLogos_yaleLogo.png';
+
+export const Data = [UCB, STANFORD, CMU, HARVARD, MIT, OXFORD, PRINCETON, YALE];
