--- src/ignitus-CoreTeam/Components/Constants.js
@@ -1,17 +1,0 @@
-export const Afelio = 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_afelio.jpeg';
-export const Divyanshu = 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_divyanshu.jpeg';
-export const Ayush = 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_ayush.jpeg';
-export const Jennifer = 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_jennifer.jpeg';
-export const Paarmita = 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_paarmita.jpeg';
-export const Rajvardhan = 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_rajvardhan.jpeg';
-export const Hafsa = 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_Hafsa.jpeg';
-export const Desire = 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_Desire.png';
-
-export const Avatar1 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-bertha.png';
-export const Avatar2 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-eugene.png';
-export const Avatar3 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-george.png';
-export const Avatar4 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-grant.png';
-export const Avatar5 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-lauren.png';
-export const Avatar6 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-melanie.png';
-export const Avatar7 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-nelson.png';
-export const Avatar8 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-robert.png';
