--- src/ignitus-CoreTeam/Components/Data.js
@@ -1,65 +1,0 @@
-import * as t from './Constants';
-
-const AVATARS = [
-  {
-    title: 'Ayush Pareek',
-    img: t.Ayush,
-    description: 'Founder, CEO',
-    linkedin: 'https://www.linkedin.com/in/ayushpareek/',
-    angellist: 'https://angel.co/ignitus-2',
-
-  },
-  {
-    title: 'Afelio Padilla',
-    img: t.Afelio,
-    description: 'COO',
-    linkedin: 'https://www.linkedin.com/in/afelio/',
-    angellist: 'https://angel.co/ignitus-2',
-  },
-  {
-    title: 'Divyanshu Rawat',
-    img: t.Divyanshu,
-    description: 'Co-founder, CTO',
-    linkedin: 'https://www.linkedin.com/in/divyanshu01/',
-    angellist: 'https://angel.co/ignitus-2',
-  },
-  {
-    title: 'Jennifer Thornton',
-    img: t.Jennifer,
-    description: 'HR & Administration',
-    linkedin: 'https://www.linkedin.com/in/jcthornton/',
-    angellist: 'https://angel.co/ignitus-2',
-  },
-  {
-    title: 'Paarmita Bhargava',
-    img: t.Paarmita,
-    description: 'Product Designer',
-    linkedin: 'https://www.linkedin.com/in/paarmitabhargava/',
-    angellist: 'https://angel.co/ignitus-2',
-
-  },
-  {
-    title: 'Desire BahBioh',
-    img: t.Desire,
-    description: 'Software Engineering',
-    linkedin: 'https://www.linkedin.com/in/ndbahbioh/',
-    angellist: 'https://angel.co/ignitus-2',
-  },
-  {
-    title: 'Rajvardhan Oak',
-    img: t.Rajvardhan,
-    description: 'Ignitus Scholar Lead',
-    linkedin: 'https://www.linkedin.com/in/rvoak/',
-    angellist: 'https://angel.co/ignitus-2',
-
-  },
-  {
-    title: 'Hafsa Nefouci',
-    img: t.Hafsa,
-    description: 'UI/UX Designer',
-    linkedin: 'https://www.linkedin.com/in/nefoucihafsa/',
-    angellist: 'https://angel.co/ignitus-2',
-
-  },
-];
-export default AVATARS;
