+++ src/ignitus-CoreTeam/Styles/style.scss
@@ -24,8 +24,8 @@
 }
 
 #team-avatar-card:hover {
-    transform: scale(1.2);
-    transition-duration: 500ms;
+    transform: scale(1.05);
+    transition-duration: 1000ms;
     cursor: pointer;
     position: absolute;
     z-index: 100;
