+++ src/ignitus-CoreTeam/constants.js
@@ -1,0 +1,63 @@
+// eslint-disable-next-line import/prefer-default-export
+export const Data = [
+  {
+    title: 'Ayush Pareek',
+    img: 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_ayush.jpeg',
+    description: 'Founder, CEO',
+    linkedin: 'https://www.linkedin.com/in/ayushpareek/',
+    angellist: 'https://angel.co/ignitus-2',
+
+  },
+  {
+    title: 'Afelio Padilla',
+    img: 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_afelio.jpeg',
+    description: 'COO',
+    linkedin: 'https://www.linkedin.com/in/afelio/',
+    angellist: 'https://angel.co/ignitus-2',
+  },
+  {
+    title: 'Divyanshu Rawat',
+    img: 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_divyanshu.jpeg',
+    description: 'Co-founder, CTO',
+    linkedin: 'https://www.linkedin.com/in/divyanshu01/',
+    angellist: 'https://angel.co/ignitus-2',
+  },
+  {
+    title: 'Jennifer Thornton',
+    img: 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_jennifer.jpeg',
+    description: 'HR & Administration',
+    linkedin: 'https://www.linkedin.com/in/jcthornton/',
+    angellist: 'https://angel.co/ignitus-2',
+  },
+  {
+    title: 'Paarmita Bhargava',
+    img: 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_paarmita.jpeg',
+    description: 'Product Designer',
+    linkedin: 'https://www.linkedin.com/in/paarmitabhargava/',
+    angellist: 'https://angel.co/ignitus-2',
+
+  },
+  {
+    title: 'Desire BahBioh',
+    img: 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_Desire.png',
+    description: 'Software Engineering',
+    linkedin: 'https://www.linkedin.com/in/ndbahbioh/',
+    angellist: 'https://angel.co/ignitus-2',
+  },
+  {
+    title: 'Rajvardhan Oak',
+    img: 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_rajvardhan.jpeg',
+    description: 'Ignitus Scholar Lead',
+    linkedin: 'https://www.linkedin.com/in/rvoak/',
+    angellist: 'https://angel.co/ignitus-2',
+
+  },
+  {
+    title: 'Hafsa Nefouci',
+    img: 'https://storage.googleapis.com/ignitus_assets/ig-team/Team_Hafsa.jpeg',
+    description: 'UI/UX Designer',
+    linkedin: 'https://www.linkedin.com/in/nefoucihafsa/',
+    angellist: 'https://angel.co/ignitus-2',
+
+  },
+];
