--- src/ignitus-Partners/Components/Constants.js
@@ -1,8 +1,0 @@
-export const WOOTECH = 'https://storage.googleapis.com/ignitus_assets/ig-partners-logo/partnersLogo_wootechLogo.png';
-export const SUGARLABS = 'https://storage.googleapis.com/ignitus_assets/ig-partners-logo/partnersLogo_sugarlabsLogo.png';
-export const SYSTERS = 'https://storage.googleapis.com/ignitus_assets/ig-partners-logo/partnersLogo_systersLogo.png';
-export const PROGATE = 'https://storage.googleapis.com/ignitus_assets/ig-partners-logo/partnersLogo_progateLogo.png';
-export const LEANIN = 'https://storage.googleapis.com/ignitus_assets/ig-partners-logo/partnersLogo_leanLogo.png';
-export const GIRLSCRIPT = 'https://storage.googleapis.com/ignitus_assets/ig-partners-logo/partnersLogo_gsLogo.png';
-
-export const Data = [WOOTECH, SUGARLABS, SYSTERS, PROGATE, LEANIN, GIRLSCRIPT];
