--- src/ignitus-Team/Components/Constants.js
@@ -1,18 +1,0 @@
-export const Afelio = 'https://storage.googleapis.com/ignitus/Team/afelio.jpeg';
-export const Divyanshu = 'https://storage.googleapis.com/ignitus/Team/divyanshu.jpeg';
-export const Ayush = 'https://storage.googleapis.com/ignitus/Team/ayush.jpeg';
-export const Jennifer = 'https://storage.googleapis.com/ignitus/Team/jennifer.jpeg';
-export const Paarmita = 'https://storage.googleapis.com/ignitus/Team/paarmita.jpeg';
-export const Rajvardhan = 'https://storage.googleapis.com/ignitus/Team/rajvardhan.jpeg';
-export const Hafsa = 'https://storage.googleapis.com/ignitus/Team/Hafsa.jpeg';
-export const Desire = 'https://storage.googleapis.com/ignitus/Team/Desire.png';
-
-
-export const Avatar1 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-bertha.png';
-export const Avatar2 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-eugene.png';
-export const Avatar3 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-george.png';
-export const Avatar4 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-grant.png';
-export const Avatar5 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-lauren.png';
-export const Avatar6 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-melanie.png';
-export const Avatar7 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-nelson.png';
-export const Avatar8 = 'https://storage.googleapis.com/ignitus_assets/ig-avatars/Team_avatar-robert.png';
