--- src/ignitus-Team/Components/Data.js
@@ -1,65 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-import * as t from './Constants';
-
-export const AVATARS = [
-  {
-    title: 'Ayush Pareek',
-    img: t.Ayush,
-    description: 'Founder, CEO',
-    linkedin: 'https://www.linkedin.com/company/igniteus?trk=ppro_cprof',
-    angellist: 'https://www.facebook.com/Ignitus.org/',
-
-  },
-  {
-    title: 'Afelio Padilla',
-    img: t.Afelio,
-    description: 'COO',
-    linkedin: 'https://www.linkedin.com/company/igniteus?trk=ppro_cprof',
-    angellist: 'https://www.facebook.com/Ignitus.org/',
-  },
-  {
-    title: 'Divyanshu Rawat',
-    img: t.Divyanshu,
-    description: 'Co-founder, CTO',
-    linkedin: 'https://www.linkedin.com/company/igniteus?trk=ppro_cprof',
-    angellist: 'https://www.facebook.com/Ignitus.org/',
-  },
-  {
-    title: 'Jennifer Thornton',
-    img: t.Jennifer,
-    description: 'HR & Administration',
-    linkedin: 'https://www.linkedin.com/company/igniteus?trk=ppro_cprof',
-    angellist: 'https://www.facebook.com/Ignitus.org/',
-  },
-  {
-    title: 'Paarmita Bhargava',
-    img: t.Paarmita,
-    description: 'Product Designer',
-    linkedin: 'https://www.linkedin.com/company/igniteus?trk=ppro_cprof',
-    facebook: 'https://www.facebook.com/Ignitus.org/',
-
-  },
-  {
-    title: 'Desire BahBioh',
-    img: t.Desire,
-    description: 'Software Engineering',
-    linkedin: 'https://www.linkedin.com/company/igniteus?trk=ppro_cprof',
-    angellist: 'https://www.facebook.com/Ignitus.org/',
-  },
-  {
-    title: 'Rajvardhan Oak',
-    img: t.Rajvardhan,
-    description: 'Ignitus Scholar Lead',
-    linkedin: 'https://www.linkedin.com/company/igniteus?trk=ppro_cprof',
-    facebook: 'https://www.facebook.com/Ignitus.org/',
-
-  },
-  {
-    title: 'Hafsa Nefouci',
-    img: t.Hafsa,
-    description: 'UI/UX Designer',
-    linkedin: 'https://www.linkedin.com/company/igniteus?trk=ppro_cprof',
-    facebook: 'https://www.facebook.com/Ignitus.org/',
-
-  },
-];
