+++ src/ignitus-Testimonial/Styles/style.scss
@@ -7,7 +7,6 @@
     text-align: center;
   }
   .testimonial-wrapper {
-    margin: 1em;
     display: flex;
     align-items: center;
     justify-content: center;
