--- src/ignitus-WhatWeDo/Styles/style.scss
@@ -1,43 +1,0 @@
-// import ignitus token variables
-@import './src/shared/styles/ignitus-tokens.scss';
-
-.ignitus-color {
-  font-weight: 400;
-  font-size: 36px;
-  color: #000E8C;
-}
-
-.card-shadow {
-  border: 0;
-
-  &:hover {
-    box-shadow: 0 4px 8px 0 #999999, 0 6px 20px 0 #999999;
-  }
-}
-
-.cardHeight {
-  height: 135px;
-  border-radius: 16px !important;
-}
-
-
-.flatten-text {
-  text-align: center;
-}
-
-.provide-card-row{
-  display: flex;
-  flex-wrap: wrap;
-  width: 70%;
-  margin-right: auto;
-  margin-left: auto;
-}
-
-.desc_card{
-  font-size: 16px;
-  font-weight: 300;
-}
-
-.card{
-    border-radius: 16px !important;
-}
